+++ NewHorizons/Components/EOTE/NHSlideCollectionContainer.cs
@@ -8,6 +8,34 @@ public class NHSlideCollectionContainer : SlideCollectionContainer
     public string[] conditionsToSet;
     public string[] persistentConditionsToSet;
 
+    [HarmonyPrefix]
+    [HarmonyPatch(typeof(SlideCollectionContainer), nameof(SlideCollectionContainer.Initialize))]
+    public static bool SlideCollectionContainer_Initialize(SlideCollectionContainer __instance)
+    {
+        if (__instance is NHSlideCollectionContainer)
+        {
+            if (__instance._initialized)
+                return false;
+            __instance.SetupReadFlags();
+            __instance.RegisterPerSlideCompletion();
+            if (__instance.streamingTexturesAvailable)
+                __instance.SetupStreaming();
+            __instance.BuildMusicRangesIndex();
+            __instance._changeSlidesAllowed = true;
+            __instance._initialized = true;
+            __instance._slideCollection.isVision = __instance._owningItem == null;
+            foreach (var factID in __instance._playWithShipLogFacts)
+            {
+                var fact = Locator.GetShipLogManager().GetFact(factID);
+                fact?.RegisterSlideCollection(__instance._slideCollection);
+            }
+            return false;
+        }
+        return true;
+    }
+
+    [HarmonyPostfix]
+    [HarmonyPatch(typeof(SlideCollectionContainer), nameof(SlideCollectionContainer.SetReadFlag))]
     public static void SlideCollectionContainer_SetReadFlag(SlideCollectionContainer __instance)
     {
         if (__instance is NHSlideCollectionContainer container)
